xingtools v1.1.2 - 预发布版本 - 新增 short 和 float 支持，新增ShortStatus和FloatStatus 注解，修复其他注解bug

+++ xingtools-core/src/main/java/top/cutexingluo/tools/common/valid/num/intstatus/IntRangeData.java
@@ -22,13 +22,13 @@ public class IntRangeData {
      *
      * <p> min value (inclusive)</p>
      */
-    protected double min;
+    protected int min;
     /**
      * 区间最大值
      *
      * <p>max value (inclusive)</p>
      */
-    protected double max;
+    protected int max;
 
     /**
      * 转化注解
